Refactor middleware into dedicated package and modular files
Decoupled middleware functionality into its own package for better organization and maintainability. Split various middleware implementations (e.g., throttler, error handling, request reducer, metrics) into separate files. Updated handler interfaces and usage to align with the new structure.
